refactor(svg): move applyConnectorRadiusFloor to connectorRadius utils
